build a real module, and update to scorep-4.0
* split the diffenent parts into different files
    * create a real package
* change the inteface to scorep.scorep_bindings.region_begin
* allow hardcoded module filtering (in scorep.cpp)
* disable tracing using import
* some things more :smile:
