+++ test/test_user_regions.py
@@ -1,10 +1,10 @@
-import scorep
+import scorep.user
 
 
 def foo():
-    scorep.user_region_begin("test_region")
+    scorep.user.region_begin("test_region")
     print("hello world")
-    scorep.user_region_end("test_region")
+    scorep.user.region_end("test_region")
 
 
 foo()
